Criando os models

+++ gestaoferias/src/main/java/gov/gestaoferias/gestaoferias/model/PublicServer.java
@@ -22,12 +22,12 @@ public class PublicServer extends Person {
     }
 
     @Override
-    public Integer getId() {
+    public Long getId() {
         return super.getId();
     }
 
     @Override
-    public void setId(Integer id) {
+    public void setId(Long id) {
         super.setId(id);
     }
